Audio Restructuring

+++ RogueEngine/Source/CameraSystem.cpp
@@ -36,7 +36,8 @@ namespace Rogue
 
 	glm::mat4 CameraSystem::GetViewMatrix(const float& parallax)
 	{
-		return glm::lookAt({ m_cameraPos.x * parallax, m_cameraPos.y * parallax, m_cameraPos.z }, glm::vec3{ m_cameraPos.x * parallax, m_cameraPos.y * parallax, m_cameraPos.z } + m_cameraFront, m_cameraUp);
+		return glm::lookAt({ static_cast<int>(m_cameraPos.x) * parallax, static_cast<int>(m_cameraPos.y) * parallax, static_cast<int>(m_cameraPos.z) },
+			glm::vec3{ static_cast<int>(m_cameraPos.x) * parallax, static_cast<int>(m_cameraPos.y) * parallax, static_cast<int>(m_cameraPos.z) } + m_cameraFront, m_cameraUp);
 	}
 
 	void CameraSystem::ResetCamera()
